# CSharp Performance: don't decrease list capacities when reusing clipper objects