--- a/CSharp/Clipper2Lib/Clipper.Offset.cs
+++ b/CSharp/Clipper2Lib/Clipper.Offset.cs
@@ -162,7 +162,7 @@ namespace Clipper2Lib
     {
       _solution.Clear();
       if (_groupList.Count == 0) return;
-      _solution.Capacity = CalcSolutionCapacity();
+      _solution.EnsureCapacity(CalcSolutionCapacity());
 
       // make sure the offset delta is significant
       if (Math.Abs(delta) < 0.5)
@@ -257,7 +257,7 @@ namespace Clipper2Lib
 
     internal static void GetMultiBounds(Paths64 paths, List<Rect64> boundsList)
     {
-      boundsList.Capacity = paths.Count;
+      boundsList.EnsureCapacity(paths.Count);
       foreach (Path64 path in paths)
       {
         if (path.Count < 1)
@@ -545,7 +545,7 @@ namespace Clipper2Lib
     {
       int cnt = path.Count;
       _normals.Clear();
-      _normals.Capacity = cnt;
+      _normals.EnsureCapacity(cnt);
 
       for (int i = 0; i < cnt - 1; i++)
         _normals.Add(GetUnitNormal(path[i], path[i + 1]));