CSharp Performance: when reusing clipper objects, currently clipper sets the internal lists capacity, even if the capacity is already sufficient. This causes C#  to drop the currently allocated array and create a new, smaller one (If clipper is not reused this is fine).

This merge request contains changes that check if the capacity is sufficient before setting it to reduce allocations.

I have conducted a benchmark (using BenchmarkDotNet) of my application that reuses clipper objects. This small change reduces allocations by 15% and also slightly improves runtimes.

| Method                         | Scenario             | solutionsToCalc | Mean    | Error    | StdDev   | Gen0       | Gen1      | Allocated |
|------------------------------- |--------------------- |---------------- |--------:|---------:|---------:|-----------:|----------:|----------:|
| NestScenarioWithStartHeuristic | 3DIPP-Instance-01259 | 4               | 6.320 s | 0.1240 s | 0.1656 s | 14000.0000 | 1000.0000 |  25.73 GB |
| NestScenarioWithStartHeuristic - no Capacity decrease | 3DIPP-Instance-01259 | 4               | 6.221 s | 0.1122 s | 0.1050 s | 12000.0000 | 1000.0000 |  21.93 GB |